+++ src/main/java/com/efs/sdk/organizationmanager/core/GlobalExceptionHandler.java
@@ -58,7 +58,7 @@ public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
 
     @ExceptionHandler(value = OrganizationmanagerException.class)
     private ResponseEntity<Object> handleOrganizationmanagerException(OrganizationmanagerException ex, WebRequest request) {
-        LOG.debug(ex.getMessage(), ex);
+        LOG.error(ex.getMessage(), ex);
         // Creating a map to hold the error details.
         Map<String, Object> body = new LinkedHashMap<>();
         body.put("timestamp", new Date());
@@ -73,7 +73,7 @@ public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
 
     @ExceptionHandler(value = AccessDeniedException.class)
     protected ResponseEntity<Object> handleAccessDeniedException(RuntimeException e, WebRequest request) {
-        LOG.debug(e.getMessage(), e);
+        LOG.error(e.getMessage(), e);
         Iterable<String> iter = request::getHeaderNames;
         Optional<String> authHeader = StreamSupport.stream(iter.spliterator(), false)
                 .filter(p -> p.equalsIgnoreCase(HttpHeaders.AUTHORIZATION)).findFirst();
@@ -90,7 +90,7 @@ public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
     @NotNull
     protected ResponseEntity<Object> handleMethodArgumentNotValid(@NotNull MethodArgumentNotValidException ex, @NotNull HttpHeaders headers,
                                                                   @NotNull HttpStatusCode status, @NotNull WebRequest request) {
-        LOG.debug(ex.getMessage(), ex);
+        LOG.error(ex.getMessage(), ex);
         List<String> errors = ex.getBindingResult()
                 .getAllErrors()
                 .stream()