+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - contributing guide
 - endpoint to get all spaces (by permissions)
 - update log level in global exception handler
+- explicitly reference DB sequences in models
 
 ### Changed
 - none