+++ CHANGELOG.md
@@ -9,26 +9,55 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-## 1.1.0 - 2023-10-05
+
+---
+
+
+## 1.1.1 - 2023-10-05
+
 
 ### Added
+
+- none
+
+### Changed
+
+- explicitly reference DB sequences in models
+- endpoint to get all spaces (by permissions)
+- update log level in global exception handler
+
+### Removed
+
+- none
+
+## 1.1.0 - 2023-10-05
+
+
+### Added
+
 - implement missing audit logs
 - fix dependency track findings
 - contributing guide
 
 ### Changed
+
 - none
 
 ### Removed
+
 - none
 
 ## 1.0.0 - 2023-09-28
 
+
 ### Added
+
 - initial release
 
 ### Changed
+
 - none
 
 ### Removed
+
 - none